Move hardhat tests to same directory
