+++ account-integrations/safe/test/hardhat/SafeCompressionPlugin.test.ts
@@ -3,15 +3,18 @@ import { getBytes, resolveProperties, ethers } from "ethers";
 import { UserOperationStruct } from "@account-abstraction/contracts";
 import { getUserOpHash } from "@account-abstraction/utils";
 import {
-  SafeECDSAFactory__factory,
-  SafeECDSAPlugin__factory,
+  AddressRegistry__factory,
+  FallbackDecompressor__factory,
+  SafeCompressionFactory__factory,
+  SafeCompressionPlugin__factory,
   SafeProxyFactory__factory,
   Safe__factory,
-} from "../../../typechain-types";
-import sendUserOpAndWait from "../utils/sendUserOpAndWait";
-import receiptOf from "../utils/receiptOf";
-import SafeSingletonFactory from "../utils/SafeSingletonFactory";
-import makeDevFaster from "../utils/makeDevFaster";
+} from "../../typechain-types";
+import sendUserOpAndWait from "./utils/sendUserOpAndWait";
+import receiptOf from "./utils/receiptOf";
+import SafeSingletonFactory from "./utils/SafeSingletonFactory";
+import makeDevFaster from "./utils/makeDevFaster";
+import sleep from "./utils/sleep";
 
 const ERC4337_TEST_ENV_VARIABLES_DEFINED =
   typeof process.env.ERC4337_TEST_BUNDLER_URL !== "undefined" &&
@@ -23,9 +26,7 @@ const BUNDLER_URL = process.env.ERC4337_TEST_BUNDLER_URL;
 const NODE_URL = process.env.ERC4337_TEST_NODE_URL;
 const MNEMONIC = process.env.MNEMONIC;
 
-const oneEther = ethers.parseEther("1");
-
-describe("SafeECDSAPlugin", () => {
+describe("SafeCompressionPlugin", () => {
   const setupTests = async () => {
     const bundlerProvider = new ethers.JsonRpcProvider(BUNDLER_URL);
     const provider = new ethers.JsonRpcProvider(NODE_URL);
@@ -33,6 +34,7 @@ describe("SafeECDSAPlugin", () => {
 
     const admin = ethers.Wallet.fromPhrase(MNEMONIC!).connect(provider);
     const userWallet = ethers.Wallet.createRandom(provider);
+
     await receiptOf(
       await admin.sendTransaction({
         to: userWallet.address,
@@ -51,9 +53,18 @@ describe("SafeECDSAPlugin", () => {
 
     const ssf = await SafeSingletonFactory.init(admin);
 
+    const safeProxyFactory = await ssf.connectOrDeploy(
+      SafeProxyFactory__factory,
+      [],
+    );
+    await safeProxyFactory.waitForDeployment();
+
+    const safeSingleton = await ssf.connectOrDeploy(Safe__factory, []);
+    await safeSingleton.waitForDeployment();
+
     return {
-      factory: await ssf.connectOrDeploy(SafeProxyFactory__factory, []),
-      singleton: await ssf.connectOrDeploy(Safe__factory, []),
+      factory: safeProxyFactory,
+      singleton: safeSingleton,
       bundlerProvider,
       provider,
       admin,
@@ -62,11 +73,14 @@ describe("SafeECDSAPlugin", () => {
     };
   };
 
-  async function setupDeployedAccount(
-    to: ethers.AddressLike,
-    value: ethers.BigNumberish,
-    data: ethers.BytesLike,
-  ) {
+  /**
+   * This test verifies a ERC4337 transaction succeeds when sent via a plugin
+   * The user operation deploys a Safe with the ERC4337 plugin and a handler
+   * and executes a transaction, thus verifying two things:
+   * 1. Deployment of the Safe with the ERC4337 plugin and handler is possible
+   * 2. Executing a transaction is possible
+   */
+  itif("should pass the ERC4337 validation", async () => {
     const {
       singleton,
       provider,
@@ -75,12 +89,13 @@ describe("SafeECDSAPlugin", () => {
       userWallet,
       entryPoints,
     } = await setupTests();
+
     const ENTRYPOINT_ADDRESS = entryPoints[0];
 
     const ssf = await SafeSingletonFactory.init(admin);
 
-    const safeECDSAFactory = await ssf.connectOrDeploy(
-      SafeECDSAFactory__factory,
+    const safeCompressionFactory = await ssf.connectOrDeploy(
+      SafeCompressionFactory__factory,
       [],
     );
 
@@ -94,37 +109,73 @@ describe("SafeECDSAPlugin", () => {
     const maxFeePerGas = `0x${feeData.maxFeePerGas.toString()}`;
     const maxPriorityFeePerGas = `0x${feeData.maxPriorityFeePerGas.toString()}`;
 
-    const owner = ethers.Wallet.createRandom();
+    const owner = ethers.Wallet.createRandom(provider);
+
+    await receiptOf(
+      admin.sendTransaction({
+        to: owner.address,
+        value: ethers.parseEther("1"),
+      }),
+    );
+
+    const addressRegistry = await ssf.connectOrDeploy(
+      AddressRegistry__factory,
+      [],
+    );
+
+    const fallbackDecompressor = await ssf.connectOrDeploy(
+      FallbackDecompressor__factory,
+      [await addressRegistry.getAddress()],
+    );
 
     const createArgs = [
       singleton,
       ENTRYPOINT_ADDRESS,
+      await fallbackDecompressor.getAddress(),
       owner.address,
       0,
-    ] satisfies Parameters<typeof safeECDSAFactory.create.staticCall>;
+    ] satisfies Parameters<typeof safeCompressionFactory.create.staticCall>;
 
-    const accountAddress = await safeECDSAFactory.create.staticCall(
+    const accountAddress = await safeCompressionFactory.create.staticCall(
       ...createArgs,
     );
 
-    await receiptOf(safeECDSAFactory.create(...createArgs));
+    await receiptOf(safeCompressionFactory.create(...createArgs));
+
+    const compressionAccount = SafeCompressionPlugin__factory.connect(
+      accountAddress,
+      userWallet,
+    );
 
     const recipient = new ethers.Wallet(
       "0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d",
     );
+
     const transferAmount = ethers.parseEther("1");
 
-    const userOpCallData =
-      SafeECDSAPlugin__factory.createInterface().encodeFunctionData(
-        "execTransaction",
-        [to, value, data],
-      );
+    const compressedActions = await fallbackDecompressor.compress(
+      [
+        {
+          to: recipient.address,
+          value: transferAmount,
+          data: "0x",
+        },
+      ],
+      [],
+    );
+    // TODO: why is this needed to prevent "nonce too low" error
+    await sleep(5000);
+
+    const userOpCallData = compressionAccount.interface.encodeFunctionData(
+      "decompressAndPerform",
+      [compressedActions],
+    );
 
     // Native tokens for the pre-fund 💸
     await receiptOf(
       admin.sendTransaction({
         to: accountAddress,
-        value: ethers.parseEther("10"),
+        value: ethers.parseEther("10"), // TODO: increasing this from 1 to 10 prevents error of balance not updating for assertion??????
       }),
     );
 
@@ -161,74 +212,13 @@ describe("SafeECDSAPlugin", () => {
       signature: userOpSignature,
     };
 
-    // Uncomment to get a detailed debug message
-    // const DEBUG_MESSAGE = `
-    //         Using entry point: ${ENTRYPOINT_ADDRESS}
-    //         Deployed Safe address: ${deployedAddress}
-    //         Module/Handler address: ${safeECDSAPluginAddress}
-    //         User operation:
-    //         ${JSON.stringify(userOperation, null, 2)}
-    //     `;
-    // console.log(DEBUG_MESSAGE);
+    const recipientBalanceBefore = await provider.getBalance(recipient.address);
 
     await sendUserOpAndWait(userOperation, ENTRYPOINT_ADDRESS, bundlerProvider);
 
-    return {
-      provider,
-      bundlerProvider,
-      entryPoint: ENTRYPOINT_ADDRESS,
-      admin,
-      userWallet,
-      accountAddress,
-    };
-  }
+    const recipientBalanceAfter = await provider.getBalance(recipient.address);
 
-  /**
-   * This test verifies a ERC4337 transaction succeeds when sent via a plugin
-   * The user operation deploys a Safe with the ERC4337 plugin and a handler
-   * and executes a transaction, thus verifying two things:
-   * 1. Deployment of the Safe with the ERC4337 plugin and handler is possible
-   * 2. Executing a transaction is possible
-   */
-  itif("should pass the ERC4337 validation", async () => {
-    const recipient = ethers.Wallet.createRandom();
-
-    const { provider } = await setupDeployedAccount(
-      recipient.address,
-      oneEther,
-      "0x",
-    );
-
-    expect(await provider.getBalance(recipient.address)).to.equal(oneEther);
-  });
-
-  itif("should not allow execTransaction from unrelated address", async () => {
-    const { accountAddress, admin, provider } = await setupDeployedAccount(
-      ethers.ZeroAddress,
-      0,
-      "0x",
-    );
-
-    const unrelatedWallet = ethers.Wallet.createRandom(provider);
-
-    await receiptOf(
-      admin.sendTransaction({
-        to: unrelatedWallet.address,
-        value: 100n * oneEther,
-      }),
-    );
-
-    const account = SafeECDSAPlugin__factory.connect(
-      accountAddress,
-      unrelatedWallet,
-    );
-
-    const recipient = ethers.Wallet.createRandom(provider);
-
-    await expect(
-      receiptOf(account.execTransaction(recipient.address, oneEther, "0x")),
-    ).to.eventually.rejected;
-
-    await expect(provider.getBalance(recipient)).to.eventually.equal(0n);
+    const expectedRecipientBalance = recipientBalanceBefore + transferAmount;
+    expect(recipientBalanceAfter).to.equal(expectedRecipientBalance);
   });
 });
